requirements.txt
fastapi
uvicorn[standard]
python-dotenv

# AI providers
anthropic>=0.40.0
google-generativeai>=0.8.0
openai>=1.0.0

# TTS dependencies
soundfile

# Flask (primary server)
Flask==3.0.0
Flask-CORS==4.0.0
gunicorn==21.2.0
sentry-sdk[flask]>=2.0.0

# Database
Flask-SQLAlchemy>=3.1.0
SQLAlchemy>=2.0.0
alembic>=1.13.0
psycopg[binary]>=3.1.18
supabase>=2.0.0

# Authentication
PyJWT>=2.8.0
google-auth>=2.0.0
requests>=2.31.0

# ElevenLabs TTS
elevenlabs>=1.0.0
boto3>=1.34.0

# Audio/Signal Processing
numpy>=1.24.0
scipy>=1.11.0
librosa>=0.10.0
